Implementing Stripe payment

+++ client/src/app/checkout/checkout-payment/checkout-payment.component.ts
@@ -1,4 +1,4 @@
-import {Component, Input, OnInit} from '@angular/core';
+import {AfterViewInit, Component, ElementRef, Input, OnDestroy, ViewChild} from '@angular/core';
 import {FormGroup} from "@angular/forms";
 import {BasketService} from "../../basket/basket.service";
 import {CheckoutService} from "../checkout.service";
@@ -7,30 +7,92 @@ import {IBasket} from "../../shared/models/basket";
 import {IOrder} from "../../shared/models/order";
 import {NavigationExtras, Router} from "@angular/router";
 
+declare var Stripe: any;
+
 @Component({
   selector: 'app-checkout-payment',
   templateUrl: './checkout-payment.component.html',
   styleUrls: ['./checkout-payment.component.scss']
 })
-export class CheckoutPaymentComponent implements OnInit {
+export class CheckoutPaymentComponent implements AfterViewInit, OnDestroy {
   @Input() checkoutForm!: FormGroup;
+
+  // Access to template reference variables in the html (for Stripe)
+  @ViewChild('cardNumber', {static: true}) cardNumberElement!: ElementRef;
+  @ViewChild('cardExpiry', {static: true}) cardExpiryElement!: ElementRef;
+  @ViewChild('cardCvc', {static: true}) cardCvcElement!: ElementRef;
+  stripe: any;
+  cardNumber: any;
+  cardExpiry: any;
+  cardCvc: any;
+  cardErrors: any;
+  cardHandler = this.onChange.bind(this);
 
   constructor(private basketService: BasketService, private checkoutService: CheckoutService,
               private toastr: ToastrService, private router: Router) {
   }
 
-  ngOnInit(): void {
+  ngAfterViewInit() {
+    this.stripe = Stripe('pk_test_51KQ0S6ChJeHkFOx2psnZ7rocGlJXBWjmesMofnvlIPZY1jRGWTfZcI9WQk0KfGM0H2tsNXR0NGnnhZSwmQOxOGO800Bvo8TDy2');
+    const elements = this.stripe.elements();
+
+    this.cardNumber = elements.create('cardNumber');
+    this.cardNumber.mount(this.cardNumberElement.nativeElement);
+    // Event listener will be looking for any changes inside the element and cardHandler will call onChange event
+    // and if there are any errors it'll populate cardErrors variable
+    this.cardNumber.addEventListener('change', this.cardHandler);
+
+    this.cardExpiry = elements.create('cardExpiry');
+    this.cardExpiry.mount(this.cardExpiryElement.nativeElement);
+    this.cardExpiry.addEventListener('change', this.cardHandler);
+
+    this.cardCvc = elements.create('cardCvc');
+    this.cardCvc.mount(this.cardCvcElement.nativeElement);
+    this.cardCvc.addEventListener('change', this.cardHandler);
+  }
+
+  // Destroying unused elements from ngAfterViewInit
+  ngOnDestroy() {
+    this.cardNumber.destroy();
+    this.cardExpiry.destroy();
+    this.cardCvc.destroy();
+  }
+
+  onChange({error}) {
+    if(error) {
+      // The error coming from Stripe
+      this.cardErrors = error.message;
+    } else {
+      this.cardErrors = null;
+    }
   }
 
   submitOrder() {
     const basket = this.basketService.getCurrentBasketValue();
     const orderToCreate = this.getOrderToCreate(basket);
     // @ts-ignore
+    // Create order and get it returned in the 'subscribe' method, then delete a local basket
     this.checkoutService.createOrder(orderToCreate).subscribe((order: IOrder) => {
       this.toastr.success('Order created successfully');
-      this.basketService.deleteLocalBasket(basket.id);
-      const navigationExtras: NavigationExtras = {state: order};
-      this.router.navigate(['checkout/success'], navigationExtras);
+      // Confirm payment intent, init payment from the customer
+      this.stripe.confirmCardPayment(basket.clientSecret, {
+        payment_method: {
+          card: this.cardNumber,
+          billing_details: {
+            name: this.checkoutForm.get('paymentForm')?.get('nameOnCard')?.value
+          }
+        }
+      }).then(result => {
+        console.log(result);
+        if(result.paymentIntent) {
+          // Perform these actions in payment intent in Stripe succeeded
+          this.basketService.deleteLocalBasket(basket.id);
+          const navigationExtras: NavigationExtras = {state: order};
+          this.router.navigate(['checkout/success'], navigationExtras);
+        } else {
+          this.toastr.error('Payment error');
+        }
+      });
     }, error => {
       this.toastr.error(error.message);
       console.log(error);
